Move conversion classes into their own package - Allows to restrict visibility of certain "internal" methods (which ConvertService offers to the ConvertTask)
